+++ data_perimeter_helper/queries/common/common_service_role_from_aws.py
@@ -31,7 +31,7 @@ logger = logging.getLogger(__name__)
 
 class common_service_role_from_aws(Query):
     """List AWS API calls made by service roles from AWS service networks.   
-You can use this query to accelerate implementation of your [**network perimeter**](https://aws.amazon.com/fr/blogs/security/establishing-a-data-perimeter-on-aws-allow-access-to-company-data-only-from-expected-networks/) controls.
+You can use this query to accelerate implementation of your [**network perimeter**](https://aws.amazon.com/blogs/security/establishing-a-data-perimeter-on-aws-allow-access-to-company-data-only-from-expected-networks/) controls.
     """  # noqa: W291
     def __init__(self, name):
         self.name = name