+++ data_perimeter_helper/queries/s3/s3_bucket_policy_identity_perimeter_ou_boundary.py
@@ -50,7 +50,7 @@ def explain_result(
 class s3_bucket_policy_identity_perimeter_ou_boundary(Query):
     """List AWS API calls made on S3 buckets in the selected account by principals that do **NOT** belong to the same organizational unit (OU) boundary, filtering out calls that align with your definition of trusted identities.
 
-You can use this query to accelerate implementation of the [**identity perimeter**](https://aws.amazon.com/fr/blogs/security/establishing-a-data-perimeter-on-aws-allow-only-trusted-identities-to-access-company-data/) controls on your S3 buckets at the organizational unit (OU) level. You can use the global condition key [aws:PrincipalOrgPaths](https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_policies_condition-keys.html#condition-keys-principalorgpaths) to limit access to your resources only to principals belonging to specific OUs.
+You can use this query to accelerate implementation of the [**identity perimeter**](https://aws.amazon.com/blogs/security/establishing-a-data-perimeter-on-aws-allow-only-trusted-identities-to-access-company-data/) controls on your S3 buckets at the organizational unit (OU) level. You can use the global condition key [aws:PrincipalOrgPaths](https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_policies_condition-keys.html#condition-keys-principalorgpaths) to limit access to your resources only to principals belonging to specific OUs.
 
 You can declare your OU boundaries in the `data perimeter helper` configuration file (`org_unit_boundary` parameter).
 OU boundaries allow you to logically regroup your accounts for analysis based on their location in your AWS organization. For example, you can declare OUs `ou-xxxxx-1111111` and `ou-xxxxx-2222222` (and all OUs contained within them) as your `production` boundary. Then you can run this query for one of the accounts in these OUs to review the API calls made by principals that do **NOT** belong to the same boundary (in this example: non-production accounts).
@@ -59,7 +59,7 @@ OU boundaries allow you to logically regroup your accounts for analysis based on
         self.name = name
         depends_on_resource_type = [
             'AWS::IAM::Role',
-            'AWS::Organizations::Account',
+            'AWS::Organizations::Tree',
             'AWS::S3::Bucket'
         ]
         super().__init__(